main.py
So far in this project, I have successfully parsed out the artist name and track name from the very nested JSON response. I added an artist_track_name() function, and this prints out all of the artists and tracks until there are no more. 

eg. Tame Impala - Is It True?
      Doja Cat - Like That (feat. Gucci Mane)
      Alex Mills - Stamina
.......
....and so on and so on....
This is the step before I put that information into a node graph.